packages/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.2.9",
]

[tool.mypy]
plugins = ['mypy_django_plugin.main']

[tool.django-stubs]
django_settings_module = "backend.settings"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.2",
    "mypy>=1.17.1",
    "ruff>=0.12.8",
]

